refactor(agent): generalize JSON field streaming
