+++ src/agent/internal/agent/speech_text.go
@@ -2,9 +2,7 @@ package agent
 
 import (
 	"encoding/json"
-	"fmt"
 	"io"
-	"strconv"
 	"strings"
 	"unicode"
 	"unicode/utf8"
@@ -147,278 +145,10 @@ func firstSpeechSentence(text string) string {
 	return ""
 }
 
-type speechTextStreamState int
-
-const (
-	speechStreamExpectObject speechTextStreamState = iota
-	speechStreamExpectKey
-	speechStreamInKey
-	speechStreamAfterKey
-	speechStreamBeforeValue
-	speechStreamInTargetString
-	speechStreamInOtherString
-	speechStreamSkipValue
-	speechStreamDone
-)
-
-type SpeechTextStreamWriter struct {
-	target io.Writer
-	field  string
-	state  speechTextStreamState
-
-	key     strings.Builder
-	pending []byte
-
-	escape        bool
-	unicodeEscape bool
-	unicodeDigits string
-	lastKey       string
-
-	depth         int
-	skipValueRoot int
-	skipString    bool
-}
+type SpeechTextStreamWriter = JSONFieldStreamWriter
 
 func NewSpeechTextStreamWriter(target io.Writer, field string) *SpeechTextStreamWriter {
-	return &SpeechTextStreamWriter{
-		target: target,
-		field:  strings.TrimSpace(field),
-		state:  speechStreamExpectObject,
-	}
-}
-
-func (w *SpeechTextStreamWriter) Write(p []byte) (int, error) {
-	if len(p) == 0 || w == nil || w.target == nil || w.field == "" {
-		return len(p), nil
-	}
-	written := len(p)
-	if len(w.pending) > 0 {
-		combined := make([]byte, 0, len(w.pending)+len(p))
-		combined = append(combined, w.pending...)
-		combined = append(combined, p...)
-		p = combined
-		w.pending = nil
-	}
-	for len(p) > 0 {
-		r, size := utf8.DecodeRune(p)
-		if r == utf8.RuneError && size == 1 {
-			if !utf8.FullRune(p) {
-				w.pending = append(w.pending[:0], p...)
-				return written, nil
-			}
-			return 0, fmt.Errorf("invalid utf8 in speech text stream")
-		}
-		if err := w.consumeRune(r); err != nil {
-			return 0, err
-		}
-		p = p[size:]
-	}
-	return written, nil
-}
-
-func (w *SpeechTextStreamWriter) consumeRune(r rune) error {
-	switch w.state {
-	case speechStreamExpectObject:
-		if unicode.IsSpace(r) {
-			return nil
-		}
-		if r == '{' {
-			w.depth = 1
-			w.state = speechStreamExpectKey
-		}
-	case speechStreamExpectKey:
-		if w.depth != 1 {
-			return nil
-		}
-		if unicode.IsSpace(r) || r == ',' {
-			return nil
-		}
-		if r == '}' {
-			w.depth = 0
-			w.state = speechStreamDone
-			return nil
-		}
-		if r == '"' {
-			w.key.Reset()
-			w.escape = false
-			w.unicodeEscape = false
-			w.unicodeDigits = ""
-			w.state = speechStreamInKey
-		}
-	case speechStreamInKey:
-		if complete, decoded, err := w.consumeJSONStringRune(r); err != nil {
-			return err
-		} else if complete {
-			w.lastKey = w.key.String()
-			w.key.Reset()
-			w.state = speechStreamAfterKey
-		} else if decoded != "" {
-			w.key.WriteString(decoded)
-		}
-	case speechStreamAfterKey:
-		if unicode.IsSpace(r) {
-			return nil
-		}
-		if r == ':' {
-			w.state = speechStreamBeforeValue
-			return nil
-		}
-		w.state = speechStreamExpectKey
-	case speechStreamBeforeValue:
-		if unicode.IsSpace(r) {
-			return nil
-		}
-		if r == '"' {
-			w.escape = false
-			w.unicodeEscape = false
-			w.unicodeDigits = ""
-			if w.lastKey == w.field {
-				w.state = speechStreamInTargetString
-			} else {
-				w.state = speechStreamInOtherString
-			}
-			return nil
-		}
-		w.startSkippingValue(r)
-	case speechStreamInTargetString:
-		if complete, decoded, err := w.consumeJSONStringRune(r); err != nil {
-			return err
-		} else if complete {
-			w.state = speechStreamDone
-		} else if decoded != "" {
-			if _, err := io.WriteString(w.target, decoded); err != nil {
-				return err
-			}
-		}
-	case speechStreamInOtherString:
-		if complete, _, err := w.consumeJSONStringRune(r); err != nil {
-			return err
-		} else if complete {
-			w.state = speechStreamExpectKey
-		}
-	case speechStreamSkipValue:
-		return w.skipValueRune(r)
-	case speechStreamDone:
-	}
-	return nil
-}
-
-func (w *SpeechTextStreamWriter) startSkippingValue(r rune) {
-	switch r {
-	case '{', '[':
-		w.skipValueRoot = w.depth
-		w.depth++
-		w.state = speechStreamSkipValue
-		w.skipString = false
-	case '"':
-		w.skipValueRoot = w.depth
-		w.escape = false
-		w.unicodeEscape = false
-		w.unicodeDigits = ""
-		w.skipString = true
-		w.state = speechStreamSkipValue
-	case '}':
-		if w.depth > 0 {
-			w.depth--
-		}
-		if w.depth == 0 {
-			w.state = speechStreamDone
-		} else {
-			w.state = speechStreamExpectKey
-		}
-	case ',':
-		w.state = speechStreamExpectKey
-	default:
-		w.skipValueRoot = w.depth
-		w.state = speechStreamSkipValue
-	}
-}
-
-func (w *SpeechTextStreamWriter) skipValueRune(r rune) error {
-	if w.skipString {
-		complete, _, err := w.consumeJSONStringRune(r)
-		if err != nil {
-			return err
-		}
-		if complete {
-			w.skipString = false
-		}
-		return nil
-	}
-	switch r {
-	case '"':
-		w.escape = false
-		w.unicodeEscape = false
-		w.unicodeDigits = ""
-		w.skipString = true
-	case '{', '[':
-		w.depth++
-	case '}', ']':
-		if w.depth > 0 {
-			w.depth--
-		}
-		if w.depth == 0 {
-			w.state = speechStreamDone
-		} else if w.depth == w.skipValueRoot {
-			w.state = speechStreamExpectKey
-		}
-	case ',':
-		if w.depth == w.skipValueRoot {
-			w.state = speechStreamExpectKey
-		}
-	}
-	return nil
-}
-
-func (w *SpeechTextStreamWriter) consumeJSONStringRune(r rune) (bool, string, error) {
-	if w.unicodeEscape {
-		w.unicodeDigits += string(r)
-		if len(w.unicodeDigits) < 4 {
-			return false, "", nil
-		}
-		value, err := strconv.ParseInt(w.unicodeDigits, 16, 32)
-		w.unicodeDigits = ""
-		w.unicodeEscape = false
-		w.escape = false
-		if err != nil {
-			return false, "", fmt.Errorf("invalid unicode escape in speech text stream: %w", err)
-		}
-		return false, string(rune(value)), nil
-	}
-
-	if w.escape {
-		w.escape = false
-		switch r {
-		case '"', '\\', '/':
-			return false, string(r), nil
-		case 'b':
-			return false, "\b", nil
-		case 'f':
-			return false, "\f", nil
-		case 'n':
-			return false, "\n", nil
-		case 'r':
-			return false, "\r", nil
-		case 't':
-			return false, "\t", nil
-		case 'u':
-			w.unicodeEscape = true
-			w.unicodeDigits = ""
-			return false, "", nil
-		default:
-			return false, "", fmt.Errorf("invalid escape in speech text stream: \\%c", r)
-		}
-	}
-
-	switch r {
-	case '\\':
-		w.escape = true
-		return false, "", nil
-	case '"':
-		return true, "", nil
-	default:
-		return false, string(r), nil
-	}
+	return NewJSONFieldStreamWriter(target, field)
 }
 
 type structuredFinalAnswer struct {
@@ -463,75 +193,7 @@ func speechStreamWriterForConfig(target io.Writer, cfg Config) io.Writer {
 		return nil
 	}
 	if cfg.VoiceSpeechSummaryEnabledOrDefault() {
-		return NewSpeechTextStreamWriter(target, "speech_text")
+		return NewJSONFieldStreamWriter(target, "speech_text")
 	}
-	return newStructuredFieldOrPlainStreamWriter(target, "output")
-}
-
-type structuredFieldOrPlainStreamWriter struct {
-	target     io.Writer
-	structured *SpeechTextStreamWriter
-	field      string
-	prefix     []byte
-	mode       int
-}
-
-const (
-	streamDetectMode = iota
-	streamStructuredMode
-	streamPlainMode
-)
-
-func newStructuredFieldOrPlainStreamWriter(target io.Writer, field string) io.Writer {
-	return &structuredFieldOrPlainStreamWriter{
-		target: target,
-		field:  strings.TrimSpace(field),
-	}
-}
-
-func (w *structuredFieldOrPlainStreamWriter) Write(p []byte) (int, error) {
-	if w == nil || w.target == nil || len(p) == 0 {
-		return len(p), nil
-	}
-	switch w.mode {
-	case streamStructuredMode:
-		_, err := w.structured.Write(p)
-		return len(p), err
-	case streamPlainMode:
-		_, err := w.target.Write(p)
-		return len(p), err
-	}
-
-	consumed := 0
-	for consumed < len(p) {
-		b := p[consumed]
-		if isJSONWhitespaceByte(b) {
-			w.prefix = append(w.prefix, b)
-			consumed++
-			continue
-		}
-		if b == '{' && w.field != "" {
-			w.mode = streamStructuredMode
-			w.structured = NewSpeechTextStreamWriter(w.target, w.field)
-			chunk := append(append([]byte{}, w.prefix...), p[consumed:]...)
-			w.prefix = nil
-			_, err := w.structured.Write(chunk)
-			return len(p), err
-		}
-		w.mode = streamPlainMode
-		chunk := append(append([]byte{}, w.prefix...), p[consumed:]...)
-		w.prefix = nil
-		_, err := w.target.Write(chunk)
-		return len(p), err
-	}
-	return len(p), nil
-}
-
-func isJSONWhitespaceByte(b byte) bool {
-	switch b {
-	case ' ', '\n', '\r', '\t':
-		return true
-	default:
-		return false
-	}
+	return NewJSONFieldOrPlainStreamWriter(target, "output")
 }
